Updated ClientManager
Added two ways of saying things globally. I hate them both. I wrote them
when I was tired, so I might rewrite later, but for now it seems like it
works. Also includes a very stupid way to check if the user is logged in
by checking if a variable is null rather than if the user is actually
logged in.