Added ChatServerListener Class
I really should have had a better name for this. But technically this
Class will store all of the @EventSystem methods that deal with the IM
Chatroom that will be used. Anyways, this Class is used for testing the
CM Class, and helping prepare for the future implementation of the
Chatroom.