Prepare To Update UserBase
Updated LoginListener to to work with the logic of the new UserBase,
which uses the ClientManager's Client class vs the UserBases's User
class to verify and store a user's information. Since the new UserBase
hasn't been finished just yet, but the idea behind it has, the logic
statements have been put into LoginListener.