Update UserBase References
Switched from PermissionsList, to PermissionsJson, because it is more
efficient to store the Permissions as a String, using JSON, than as
Java's Serialization and storing it as a byte array. Also added the new
way to set a user's information, because the User class is obsolete in
most cases, and is no longer needed to verify a user; no need to create
extra object when I can directly edit the Client class.